[tornado] provide TornadoTestCase for easy testing
